cleaned up platinumEyes.h
-moved all function declarations to their respective header files, that are in their respective folders, which made platinumEyes.h much cleaner.
