+++ config/configUtilities.cpp
@@ -1,6 +1,6 @@
-#include "../platinumEyes.h"
+#include "config.h"
 #include <direct.h>   
-
+#include "../platinumEyes.h"
