+++ gui/settingsMenu.cpp
@@ -1,4 +1,4 @@
-#include "../platinumEyes.h"
+#include "gui.h"
 #include <functional>
 #include <imgui_stdlib.h>
 #include "settingsChanged.h"
