+++ relics/relicDatabase.cpp
@@ -1,4 +1,4 @@
-#include "../platinumEyes.h"
+#include "relics.h"
 #include <regex>
