+++ utilities/utilities.h
@@ -1,0 +1,7 @@
+#pragma once
+#include "../platinumEyes.h"
+
+
+
+std::pair<int, int>stringToIntPair(std::string s);
+
